Migrated db/pg-query.js to typescript.
